Vongtrung

+++ src/main/java/dao/ReceiveDeliveryRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ReceiveDeliveryRepository extends JpaRepository<ReceiveDelivery, Integer> {
+public interface ReceiveDeliveryRepository extends JpaRepository<ReceiveDelivery, Long> {
     @Query(value = "select o from ReceiveDelivery o where o.shipper.id =:shipperCode")
     public List <ReceiveDelivery> findAllForshipper(@Param("shipperCode")Long shipperCode);
